Revert "remove custom serialization for optimization from PoolConfig, VaultState, AmplificationState  (#605)"
This reverts commit 87091b7e2300c63385a17ec460a035de400f4d7e.

+++ pkg/interfaces/contracts/test/IVaultMainMock.sol
@@ -26,7 +26,7 @@ interface IVaultMainMock {
     function manualRegisterPoolAtTimestamp(
         address pool,
         IERC20[] memory tokens,
-        uint32 timestamp,
+        uint256 timestamp,
         PoolRoleAccounts memory roleAccounts
     ) external;
 
@@ -36,7 +36,7 @@ interface IVaultMainMock {
 
     function manualSetPoolPaused(address, bool) external;
 
-    function manualSetPoolPauseWindowEndTime(address, uint32) external;
+    function manualSetPoolPauseWindowEndTime(address, uint256) external;
 
     function manualSetVaultPaused(bool) external;
