+++ pkg/interfaces/contracts/vault/IVaultExtension.sol
@@ -77,7 +77,7 @@ interface IVaultExtension {
         address pool,
         TokenConfig[] memory tokenConfig,
         uint256 swapFeePercentage,
-        uint32 pauseWindowEndTime,
+        uint256 pauseWindowEndTime,
         PoolRoleAccounts calldata roleAccounts,
         address poolHooksContract,
         LiquidityManagement calldata liquidityManagement
@@ -249,7 +249,7 @@ interface IVaultExtension {
      * @return poolBufferPeriodEndTime The timestamp after which the Pool unpauses itself (if paused)
      * @return pauseManager The pause manager, or the zero address
      */
-    function getPoolPausedState(address pool) external view returns (bool, uint32, uint32, address);
+    function getPoolPausedState(address pool) external view returns (bool, uint256, uint256, address);
 
     /*******************************************************************************
                                    Fees
