+++ pkg/pool-weighted/contracts/WeightedPoolFactory.sol
@@ -25,7 +25,7 @@ contract WeightedPoolFactory is IPoolVersion, BasePoolFactory, Version {
 
     constructor(
         IVault vault,
-        uint32 pauseWindowDuration,
+        uint256 pauseWindowDuration,
         string memory factoryVersion,
         string memory poolVersion
     ) BasePoolFactory(vault, pauseWindowDuration, type(WeightedPool).creationCode) Version(factoryVersion) {
