+++ pkg/pool-weighted/test/foundry/WeightedPool8020Factory.t.sol
@@ -49,7 +49,7 @@ contract WeightedPool8020FactoryTest is Test {
     }
 
     function testFactoryPausedState() public {
-        uint32 pauseWindowDuration = factory.getPauseWindowDuration();
+        uint256 pauseWindowDuration = factory.getPauseWindowDuration();
         assertEq(pauseWindowDuration, 365 days);
     }
