+++ pkg/pool-weighted/test/foundry/WeightedPoolFactory.t.sol
@@ -20,7 +20,7 @@ contract WeightedPoolFactoryTest is Test {
     }
 
     function testFactoryPausedState() public {
-        uint32 pauseWindowDuration = factory.getPauseWindowDuration();
+        uint256 pauseWindowDuration = factory.getPauseWindowDuration();
         assertEq(pauseWindowDuration, 365 days);
     }
 }
