+++ pkg/vault/contracts/VaultFactory.sol
@@ -30,8 +30,8 @@ contract VaultFactory is Authentication {
     bool public isDisabled;
 
     IAuthorizer private immutable _authorizer;
-    uint32 private immutable _pauseWindowDuration;
-    uint32 private immutable _bufferPeriodDuration;
+    uint256 private immutable _pauseWindowDuration;
+    uint256 private immutable _bufferPeriodDuration;
     address private immutable _deployer;
 
     bytes private _creationCode;
@@ -40,8 +40,8 @@ contract VaultFactory is Authentication {
 
     constructor(
         IAuthorizer authorizer,
-        uint32 pauseWindowDuration,
-        uint32 bufferPeriodDuration
+        uint256 pauseWindowDuration,
+        uint256 bufferPeriodDuration
     ) Authentication(bytes32(uint256(uint160(address(this))))) {
         _deployer = msg.sender;
         _creationCode = type(Vault).creationCode;
