+++ pkg/vault/test/foundry/unit/PoolAndVaultPaused.t.sol
@@ -15,7 +15,7 @@ import { BaseVaultTest } from "../utils/BaseVaultTest.sol";
 contract PoolAndVaultPausedTest is BaseVaultTest {
     // A number that is much smaller than the vault pause buffer end time, so we can play with
     // pool and vault pause windows.
-    uint32 private constant _FIXED_POOL_PAUSE_END_TIME = 1e5;
+    uint256 private constant _FIXED_POOL_PAUSE_END_TIME = 1e5;
     uint256 private _vaultBufferPeriodEndTimeTest;
 
     function setUp() public virtual override {
@@ -134,7 +134,7 @@ contract PoolAndVaultPausedTest is BaseVaultTest {
 
     // Returns the correct block.timestamp to consider the pool unpaused
     function _getTimeAfterPoolPauseBufferPeriod() private view returns (uint256) {
-        uint32 bufferPeriodDuration = vault.getBufferPeriodDuration();
+        uint256 bufferPeriodDuration = vault.getBufferPeriodDuration();
         return _FIXED_POOL_PAUSE_END_TIME + bufferPeriodDuration + 1;
     }
 }
